Optimized text preprocessor
The preprocessor now removes tokens with numeric characters and symbols. Removal of stopwords is made optional since googles word2vec model has some stopwords, and it's a very slow process. Methods that did nothing but itterate the tokens and perform it's operations have been concatenated, such that there are fewer itterations to be performed. Lastly the word2vec option was removed.